+++ internal/provider/model_iosxe_policy_map.go
@@ -88,6 +88,7 @@ type PolicyMapClassesActions struct {
 	PolicePirBe                          types.Int64  `tfsdk:"police_pir_be"`
 	PoliceCirConformTransmit             types.Bool   `tfsdk:"police_cir_conform_transmit"`
 	PoliceCirExceedDrop                  types.Bool   `tfsdk:"police_cir_exceed_drop"`
+	PoliceCirExceedTransmit              types.Bool   `tfsdk:"police_cir_exceed_transmit"`
 	PoliceRatePercent                    types.Int64  `tfsdk:"police_rate_percent"`
 	QueueBuffersRatio                    types.Int64  `tfsdk:"queue_buffers_ratio"`
 	SetDscp                              types.String `tfsdk:"set_dscp"`
@@ -141,6 +142,7 @@ type PolicyMapClassesActionsData struct {
 	PolicePirBe                          types.Int64  `tfsdk:"police_pir_be"`
 	PoliceCirConformTransmit             types.Bool   `tfsdk:"police_cir_conform_transmit"`
 	PoliceCirExceedDrop                  types.Bool   `tfsdk:"police_cir_exceed_drop"`
+	PoliceCirExceedTransmit              types.Bool   `tfsdk:"police_cir_exceed_transmit"`
 	PoliceRatePercent                    types.Int64  `tfsdk:"police_rate_percent"`
 	QueueBuffersRatio                    types.Int64  `tfsdk:"queue_buffers_ratio"`
 	SetDscp                              types.String `tfsdk:"set_dscp"`
@@ -325,6 +327,11 @@ func (data PolicyMap) toBody(ctx context.Context, config PolicyMap) string {
 							body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"class"+"."+strconv.Itoa(index)+"."+"action-list"+"."+strconv.Itoa(cindex)+"."+"police-policy-map.police.actions.exceed-drop.exceed-action.drop", map[string]string{})
 						}
 					}
+					if !citem.PoliceCirExceedTransmit.IsNull() && !citem.PoliceCirExceedTransmit.IsUnknown() {
+						if citem.PoliceCirExceedTransmit.ValueBool() {
+							body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"class"+"."+strconv.Itoa(index)+"."+"action-list"+"."+strconv.Itoa(cindex)+"."+"police-policy-map.police.actions.exceed-transmit.exceed-action.transmit", map[string]string{})
+						}
+					}
 					if !citem.PoliceRatePercent.IsNull() && !citem.PoliceRatePercent.IsUnknown() {
 						body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"class"+"."+strconv.Itoa(index)+"."+"action-list"+"."+strconv.Itoa(cindex)+"."+"police-rate-percent.police.rate.percent.percentage", strconv.FormatInt(citem.PoliceRatePercent.ValueInt64(), 10))
 					}
@@ -500,6 +507,13 @@ func (data PolicyMap) toBodyXML(ctx context.Context, config PolicyMap) string {
 							ccBody = helpers.SetFromXPath(ccBody, "police-policy-map/police/actions/exceed-drop/exceed-action/drop", "")
 						} else {
 							ccBody = helpers.RemoveFromXPath(ccBody, "police-policy-map/police/actions/exceed-drop/exceed-action/drop")
+						}
+					}
+					if !citem.PoliceCirExceedTransmit.IsNull() && !citem.PoliceCirExceedTransmit.IsUnknown() {
+						if citem.PoliceCirExceedTransmit.ValueBool() {
+							ccBody = helpers.SetFromXPath(ccBody, "police-policy-map/police/actions/exceed-transmit/exceed-action/transmit", "")
+						} else {
+							ccBody = helpers.RemoveFromXPath(ccBody, "police-policy-map/police/actions/exceed-transmit/exceed-action/transmit")
 						}
 					}
 					if !citem.PoliceRatePercent.IsNull() && !citem.PoliceRatePercent.IsUnknown() {
@@ -804,6 +818,15 @@ func (data *PolicyMap) updateFromBody(ctx context.Context, res gjson.Result) {
 			} else {
 				data.Classes[i].Actions[ci].PoliceCirExceedDrop = types.BoolNull()
 			}
+			if value := cr.Get("police-policy-map.police.actions.exceed-transmit.exceed-action.transmit"); !data.Classes[i].Actions[ci].PoliceCirExceedTransmit.IsNull() {
+				if value.Exists() {
+					data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolValue(true)
+				} else {
+					data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolValue(false)
+				}
+			} else {
+				data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolNull()
+			}
 			if value := cr.Get("police-rate-percent.police.rate.percent.percentage"); value.Exists() && !data.Classes[i].Actions[ci].PoliceRatePercent.IsNull() {
 				data.Classes[i].Actions[ci].PoliceRatePercent = types.Int64Value(value.Int())
 			} else {
@@ -1101,6 +1124,15 @@ func (data *PolicyMap) updateFromBodyXML(ctx context.Context, res xmldot.Result)
 			} else {
 				data.Classes[i].Actions[ci].PoliceCirExceedDrop = types.BoolNull()
 			}
+			if value := helpers.GetFromXPath(cr, "police-policy-map/police/actions/exceed-transmit/exceed-action/transmit"); !data.Classes[i].Actions[ci].PoliceCirExceedTransmit.IsNull() {
+				if value.Exists() {
+					data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolValue(true)
+				} else {
+					data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolValue(false)
+				}
+			} else {
+				data.Classes[i].Actions[ci].PoliceCirExceedTransmit = types.BoolNull()
+			}
 			if value := helpers.GetFromXPath(cr, "police-rate-percent/police/rate/percent/percentage"); value.Exists() && !data.Classes[i].Actions[ci].PoliceRatePercent.IsNull() {
 				data.Classes[i].Actions[ci].PoliceRatePercent = types.Int64Value(value.Int())
 			} else {
@@ -1269,6 +1301,11 @@ func (data *PolicyMap) fromBody(ctx context.Context, res gjson.Result) {
 					} else {
 						cItem.PoliceCirExceedDrop = types.BoolValue(false)
 					}
+					if ccValue := cv.Get("police-policy-map.police.actions.exceed-transmit.exceed-action.transmit"); ccValue.Exists() {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(true)
+					} else {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(false)
+					}
 					if ccValue := cv.Get("police-rate-percent.police.rate.percent.percentage"); ccValue.Exists() {
 						cItem.PoliceRatePercent = types.Int64Value(ccValue.Int())
 					}
@@ -1435,6 +1472,11 @@ func (data *PolicyMapData) fromBody(ctx context.Context, res gjson.Result) {
 					} else {
 						cItem.PoliceCirExceedDrop = types.BoolValue(false)
 					}
+					if ccValue := cv.Get("police-policy-map.police.actions.exceed-transmit.exceed-action.transmit"); ccValue.Exists() {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(true)
+					} else {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(false)
+					}
 					if ccValue := cv.Get("police-rate-percent.police.rate.percent.percentage"); ccValue.Exists() {
 						cItem.PoliceRatePercent = types.Int64Value(ccValue.Int())
 					}
@@ -1596,6 +1638,11 @@ func (data *PolicyMap) fromBodyXML(ctx context.Context, res xmldot.Result) {
 						cItem.PoliceCirExceedDrop = types.BoolValue(true)
 					} else {
 						cItem.PoliceCirExceedDrop = types.BoolValue(false)
+					}
+					if ccValue := helpers.GetFromXPath(cv, "police-policy-map/police/actions/exceed-transmit/exceed-action/transmit"); ccValue.Exists() {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(true)
+					} else {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(false)
 					}
 					if ccValue := helpers.GetFromXPath(cv, "police-rate-percent/police/rate/percent/percentage"); ccValue.Exists() {
 						cItem.PoliceRatePercent = types.Int64Value(ccValue.Int())
@@ -1759,6 +1806,11 @@ func (data *PolicyMapData) fromBodyXML(ctx context.Context, res xmldot.Result) {
 					} else {
 						cItem.PoliceCirExceedDrop = types.BoolValue(false)
 					}
+					if ccValue := helpers.GetFromXPath(cv, "police-policy-map/police/actions/exceed-transmit/exceed-action/transmit"); ccValue.Exists() {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(true)
+					} else {
+						cItem.PoliceCirExceedTransmit = types.BoolValue(false)
+					}
 					if ccValue := helpers.GetFromXPath(cv, "police-rate-percent/police/rate/percent/percentage"); ccValue.Exists() {
 						cItem.PoliceRatePercent = types.Int64Value(ccValue.Int())
 					}
@@ -1834,6 +1886,9 @@ func (data *PolicyMap) getDeletedItems(ctx context.Context, state PolicyMap) []s
 							}
 							if !state.Classes[i].Actions[ci].PoliceRatePercent.IsNull() && data.Classes[j].Actions[cj].PoliceRatePercent.IsNull() {
 								deletedItems = append(deletedItems, fmt.Sprintf("%v/class=%v/action-list=%v/police-rate-percent/police/rate/percent/percentage", state.getPath(), strings.Join(stateKeyValues[:], ","), strings.Join(cstateKeyValues[:], ",")))
+							}
+							if !state.Classes[i].Actions[ci].PoliceCirExceedTransmit.IsNull() && data.Classes[j].Actions[cj].PoliceCirExceedTransmit.IsNull() {
+								deletedItems = append(deletedItems, fmt.Sprintf("%v/class=%v/action-list=%v/police-policy-map/police/actions/exceed-transmit/exceed-action/transmit", state.getPath(), strings.Join(stateKeyValues[:], ","), strings.Join(cstateKeyValues[:], ",")))
 							}
 							if !state.Classes[i].Actions[ci].PoliceCirExceedDrop.IsNull() && data.Classes[j].Actions[cj].PoliceCirExceedDrop.IsNull() {
 								deletedItems = append(deletedItems, fmt.Sprintf("%v/class=%v/action-list=%v/police-policy-map/police/actions/exceed-drop/exceed-action/drop", state.getPath(), strings.Join(stateKeyValues[:], ","), strings.Join(cstateKeyValues[:], ",")))
@@ -2022,6 +2077,9 @@ func (data *PolicyMap) addDeletedItemsXML(ctx context.Context, state PolicyMap, 
 							if !state.Classes[i].Actions[ci].PoliceRatePercent.IsNull() && data.Classes[j].Actions[cj].PoliceRatePercent.IsNull() {
 								b = helpers.RemoveFromXPath(b, fmt.Sprintf(state.getXPath()+"/class%v/action-list%v/police-rate-percent/police/rate/percent/percentage", predicates, cpredicates))
 							}
+							if !state.Classes[i].Actions[ci].PoliceCirExceedTransmit.IsNull() && data.Classes[j].Actions[cj].PoliceCirExceedTransmit.IsNull() {
+								b = helpers.RemoveFromXPath(b, fmt.Sprintf(state.getXPath()+"/class%v/action-list%v/police-policy-map/police/actions/exceed-transmit/exceed-action/transmit", predicates, cpredicates))
+							}
 							if !state.Classes[i].Actions[ci].PoliceCirExceedDrop.IsNull() && data.Classes[j].Actions[cj].PoliceCirExceedDrop.IsNull() {
 								b = helpers.RemoveFromXPath(b, fmt.Sprintf(state.getXPath()+"/class%v/action-list%v/police-policy-map/police/actions/exceed-drop/exceed-action/drop", predicates, cpredicates))
 							}
@@ -2158,6 +2216,9 @@ func (data *PolicyMap) getEmptyLeafsDelete(ctx context.Context) []string {
 
 		for ci := range data.Classes[i].Actions {
 			ckeyValues := [...]string{data.Classes[i].Actions[ci].Type.ValueString()}
+			if !data.Classes[i].Actions[ci].PoliceCirExceedTransmit.IsNull() && !data.Classes[i].Actions[ci].PoliceCirExceedTransmit.ValueBool() {
+				emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/class=%v/action-list=%v/police-policy-map/police/actions/exceed-transmit/exceed-action/transmit", data.getPath(), strings.Join(keyValues[:], ","), strings.Join(ckeyValues[:], ",")))
+			}
 			if !data.Classes[i].Actions[ci].PoliceCirExceedDrop.IsNull() && !data.Classes[i].Actions[ci].PoliceCirExceedDrop.ValueBool() {
 				emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/class=%v/action-list=%v/police-policy-map/police/actions/exceed-drop/exceed-action/drop", data.getPath(), strings.Join(keyValues[:], ","), strings.Join(ckeyValues[:], ",")))
 			}